refactored enemy controller
