+++ Assets/Scripts/Enemy/EnemyTypes/TankType.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4c6174356ae63594fa8d59bf9534ea17
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
