+++ Assets/Scripts/Enemy/EnemyView.cs
@@ -5,10 +5,5 @@ public class EnemyView : MonoBehaviour
 {
     [SerializeField] private VisualEffect _footStepVFX;
 
-    public void Initialize(VFXScriptableObject footStepVFXSO)
-    {
-        _footStepVFX = footStepVFXSO.VisualEffect;
-    }
-
     public void BurstFootStep() => _footStepVFX.Play();
 }
