+++ Assets/Scripts/Main/GameService.cs
@@ -17,14 +17,8 @@ public class GameService : MonoBehaviour
     [SerializeField] private NavMeshAgent _enemyAgent;
     [SerializeField] private Animator _enemyAnimator;
 
-    [Header("VFX Properties")]
-    [SerializeField] private EnemyView _enemyVFXController;
-    [SerializeField] private PlayerView _playerVFXController;
-    [SerializeField] private List<VFXScriptableObject> _vFXSOList;
-
     private PlayerService _playerService;
     private EnemyService _enemyService;
-    private VFXService _vFXService;
 
     private void Start()
     {
