+++ Assets/Scripts/Miscellaneous/PlayerDetectionCollider.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e03eae89c63efa343b18c3d5b1fd8f0b
+guid: 6f7169f1358416942b205c3677214caa
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
