--- Assets/Scripts/VFX/VFXOwner.cs
@@ -1,5 +1,0 @@
-﻿public enum VFXOwner
-{
-    Player,
-    Enemy
-}
