--- Assets/Scripts/VFX/VFXScriptableObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1704436b407429e4f90119aa66b246c8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
