--- Assets/Scripts/VFX/VFXService.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fbded73bbb314ed4aae8ff24f075eb15
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
