+++ UserSettings/Layouts/default-2022.dwlt
@@ -19,7 +19,7 @@ MonoBehaviour:
     width: 2560
     height: 1357
   m_ShowMode: 4
-  m_Title: Hierarchy
+  m_Title: Game
   m_RootView: {fileID: 2}
   m_MinSize: {x: 875, y: 300}
   m_MaxSize: {x: 10000, y: 10000}
@@ -120,7 +120,7 @@ MonoBehaviour:
   m_MinSize: {x: 300, y: 100}
   m_MaxSize: {x: 24288, y: 16192}
   vertical: 0
-  controlID: 118
+  controlID: 159
 --- !u!114 &6
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -145,7 +145,7 @@ MonoBehaviour:
   m_MinSize: {x: 100, y: 100}
   m_MaxSize: {x: 8096, y: 16192}
   vertical: 1
-  controlID: 911
+  controlID: 59
 --- !u!114 &7
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -156,7 +156,7 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 1
   m_Script: {fileID: 12006, guid: 0000000000000000e000000000000000, type: 0}
-  m_Name: AnimatorControllerTool
+  m_Name: SceneView
   m_EditorClassIdentifier: 
   m_Children: []
   m_Position:
@@ -165,14 +165,14 @@ MonoBehaviour:
     y: 0
     width: 1564
     height: 656
-  m_MinSize: {x: 100, y: 100}
-  m_MaxSize: {x: 4000, y: 4000}
-  m_ActualView: {fileID: 16}
+  m_MinSize: {x: 201, y: 221}
+  m_MaxSize: {x: 4001, y: 4021}
+  m_ActualView: {fileID: 15}
   m_Panes:
   - {fileID: 15}
   - {fileID: 16}
-  m_Selected: 1
-  m_LastSelected: 0
+  m_Selected: 0
+  m_LastSelected: 1
 --- !u!114 &8
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -226,7 +226,7 @@ MonoBehaviour:
   m_MinSize: {x: 100, y: 100}
   m_MaxSize: {x: 8096, y: 16192}
   vertical: 1
-  controlID: 53
+  controlID: 83
 --- !u!114 &10
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -859,9 +859,9 @@ MonoBehaviour:
   m_PlayAudio: 0
   m_AudioPlay: 0
   m_Position:
-    m_Target: {x: -0.23863225, y: 6.669196, z: 31.442146}
+    m_Target: {x: 2.6354733, y: 7.229459, z: 24.14461}
     speed: 2
-    m_Value: {x: -0.23863225, y: 6.669196, z: 31.442146}
+    m_Value: {x: 2.6354733, y: 7.229459, z: 24.14461}
   m_RenderMode: 0
   m_CameraMode:
     drawMode: 0
@@ -907,9 +907,9 @@ MonoBehaviour:
     m_GridAxis: 1
     m_gridOpacity: 0.5
   m_Rotation:
-    m_Target: {x: 0.29629964, y: -0.07725836, z: 0.024143731, w: 0.95177704}
+    m_Target: {x: 0.11018099, y: -0.12640145, z: 0.014200873, w: 0.9858585}
     speed: 2
-    m_Value: {x: 0.29626632, y: -0.077249676, z: 0.024141017, w: 0.95167005}
+    m_Value: {x: 0.110168, y: -0.12638655, z: 0.014199198, w: 0.9857423}
   m_Size:
     m_Target: 2.885219
     speed: 2
@@ -977,14 +977,14 @@ MonoBehaviour:
     - {fileID: 4400659718044310736, guid: 1259ef57b46a26d4ca7b0be3edddfb40, type: 2}
     - {fileID: 7325472727431645091, guid: 1259ef57b46a26d4ca7b0be3edddfb40, type: 2}
     m_ValueSerializationHelper:
-    - e00: 1.6141026
+    - e00: 1.1399499
       e01: 0
       e02: 0
-      e03: 337.8205
+      e03: 393.89743
       e10: 0
-      e11: 1.6141026
+      e11: 1.1399499
       e12: 0
-      e13: -56.295532
+      e13: 62.593353
       e20: 0
       e21: 0
       e22: 1
@@ -1095,7 +1095,7 @@ MonoBehaviour:
     m_OverlaysVisible: 1
   m_LockTracker:
     m_IsLocked: 0
-  m_LastSelectedObjectID: 29202
+  m_LastSelectedObjectID: 28996
 --- !u!114 &19
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -1176,9 +1176,9 @@ MonoBehaviour:
   m_SceneHierarchy:
     m_TreeViewState:
       scrollPos: {x: 0, y: 0}
-      m_SelectedIDs: 
-      m_LastClickedID: 0
-      m_ExpandedIDs: 6af9ffff0efbffff0e710000
+      m_SelectedIDs: 44710000
+      m_LastClickedID: 28996
+      m_ExpandedIDs: 6af9ffff0efbfffffa7000000e71000012720000227200003e720000
       m_RenameOverlay:
         m_UserAcceptedRename: 0
         m_Name: 
@@ -1246,8 +1246,7 @@ MonoBehaviour:
     m_ShowAllHits: 0
     m_SkipHidden: 0
     m_SearchArea: 1
-    m_Folders:
-    - Assets/Scripts/Enemy
+    m_Folders: []
     m_Globs: []
     m_OriginalText: 
     m_ImportLogFlags: 0
@@ -1263,7 +1262,7 @@ MonoBehaviour:
     scrollPos: {x: 0, y: 0}
     m_SelectedIDs: b8770000
     m_LastClickedID: 30648
-    m_ExpandedIDs: 000000001aa300001ca300001ea3000020a3000022a3000024a3000026a3000028a30000
+    m_ExpandedIDs: 000000001aa300001ca300001ea3000020a3000022a3000024a3000026a3000028a300002aa300002ca300002ea3000030a3000032a30000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -1291,7 +1290,7 @@ MonoBehaviour:
     scrollPos: {x: 0, y: 0}
     m_SelectedIDs: 
     m_LastClickedID: 0
-    m_ExpandedIDs: ffffffff000000001aa300001ca300001ea3000020a3000022a3000024a3000026a3000028a3000074a5000076a50000eca50000eea50000f0a50000
+    m_ExpandedIDs: ffffffff000000001aa300001ca300001ea3000020a3000022a3000024a3000026a3000028a3000070a6000074a90000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -1307,7 +1306,7 @@ MonoBehaviour:
       m_IsRenaming: 0
       m_OriginalEventType: 11
       m_IsRenamingFilename: 1
-      m_ClientGUIView: {fileID: 11}
+      m_ClientGUIView: {fileID: 7}
     m_SearchString: 
     m_CreateAssetUtility:
       m_EndAction: {fileID: 0}
